[CHORE] 리뷰 등록 수정 (#119)
